% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_formation_head.R
\name{estimate_formation_head}
\alias{estimate_formation_head}
\title{Estimate formation heads given a set of transmissivity values and blended
head measurements.}
\usage{
estimate_formation_head(transmissivity, blended)
}
\arguments{
\item{transmissivity}{the transmissivity distribution for each interval}

\item{blended}{the blended head value}
}
\value{
a single value of the blended head
}
\description{
`estimate_formation_head` takes a vector of transmissivity and a vector of
  associated blended head measurements. The formation head is then
  determined as each interval is added. The transmissivity and blended head
  values should be ordered so that the monitoring interval is incrementally
  increased with each value. An example application would be the FLUTe
  reverse head methodology
}
\details{
`estimate_formation_head` calculates the formation head using the
  following formula:

\deqn{^fh_i = \frac{\sum\limits_{i=1}^{i} \Delta T_i (^bh_i - ^fh_i)}{T_i} + ^bh_1}

where:

 * \eqn{h_{blended}} is the blended head in the well
 * \eqn{\Delta T_i} is the difference in transmissivity between increasingly
   larger interval.
 * \eqn{^fh_i} is the formation head of the newly exposed interval
 * \eqn{^bh_i} is the blended head of the increasingly larger interval
 * \eqn{^bh_1} is the blended head of the first interval
 * \eqn{i} is the interval number
}
\examples{
transmissivity <- abs(rnorm(10, sd = 100))       # random T distribution
cumulative_transmissivity <- cumsum(transmissivity)
head <- sort(rnorm(10))                          # decreasing head with depth

formation_head <- estimate_formation_head(cumulative_transmissivity, head)
estimate_blended_head(transmissivity[1:3], formation_head[1:3])
}
\references{
Keller, C. (2017). A New Rapid Method for Measuring the Vertical
  Head Profile. Groundwater, 55(2), 244-254.
}
\seealso{
Other {blended head methods}: 
\code{\link{blended_head_change}()},
\code{\link{estimate_blended_head}()},
\code{\link{estimate_missing}()},
\code{\link{plot_blended}()}
}
\concept{{blended head methods}}
