% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_missing.R
\name{estimate_missing}
\alias{estimate_missing}
\title{Calculate the transmissivity weighted blended head}
\usage{
estimate_missing(blended = NA_real_, transmissivity, head)
}
\arguments{
\item{blended}{the blended head value}

\item{transmissivity}{the transmissivity distribution for each interval}

\item{head}{the head distribution for each interval}
}
\value{
an estimation for the missing value
}
\description{
`estimate_missing` takes a single blended head value and a vector of
  transmissivity and a vector of hydraulic head and estimates a single
  missing value of transmissivity or head.  This can be useful if you know
  all the formation heads but are missing the transmissivity distribution, or
  if you know the transmissivity distribution but not all the head values.
  Only a single missing value can be estimated at this point.
}
\examples{

# determine the relative magnitude of transmissivity for a zone with known
# head

partial_transmissivity <- 1
missing_transmissivity <-
  estimate_missing(blended = 0,
                   transmissivity = c(partial_transmissivity, NA_real_),
                   head = c(0.4, -1.9))
missing_transmissivity / (partial_transmissivity + missing_transmissivity)

}
\seealso{
Other {blended head methods}: 
\code{\link{blended_head_change}()},
\code{\link{estimate_blended_head}()},
\code{\link{estimate_formation_head}()},
\code{\link{plot_blended}()}
}
\concept{{blended head methods}}
