% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blended.R
\name{plot_blended}
\alias{plot_blended}
\title{Plot the head and transmissivity with the blended head}
\usage{
plot_blended(transmissivity, head)
}
\arguments{
\item{transmissivity}{the transmissivity distribution for each interval}

\item{head}{the head distribution for each interval}
}
\value{
a ggplot of the transmissivity and head
}
\description{
`plot_blended` takes a vector of transmissivity and a vector of
hydraulic head and calculates the expected blended head for a
cross-connected well and returns a ggplot2 object. The length units for
transmissivity and hydraulic head should be the same.
}
\details{
The plotting function usese `estimate_blended_head` to calculate
  the blended head value.  The blended head is plotted with a dashed line.
}
\examples{
transmissivity <- abs(rnorm(10, sd = 100))       # random T distribution
head <- sort(rnorm(10))                          # decreasing head with depth

plot_blended(transmissivity, head)
}
\references{
Sokol, D. (1963). Position and fluctuations of water level in
  wells perforated in more than one aquifer. Journal of Geophysical Research,
  68(4), 1079-1080.
}
\seealso{
Other {blended head methods}: 
\code{\link{blended_head_change}()},
\code{\link{estimate_blended_head}()},
\code{\link{estimate_formation_head}()},
\code{\link{estimate_missing}()}
}
\concept{{blended head methods}}
