% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sokol-package.R
\docType{package}
\name{sokol}
\alias{sokol}
\alias{_PACKAGE}
\alias{sokol-package}
\title{Blended water level analysis for wells}
\description{
This package provides a few method to assist in the
  interpretation of wells that cross-connect multiple hydrogeologic
  units with variable hydraulic head and transmissivity profiles.
}
\author{
\strong{Maintainer}: Jonathan Kennel \email{jkennel@uoguelph.ca} [translator]

Other contributors:
\itemize{
  \item Beth Parker \email{bparker@uoguelph.ca} [thesis advisor]
}

}
\keyword{internal}
